Menus: split hardware config and properties from Tools menu to new Config menu.
